Update require function to work with select lists
In the require function, add a check to see if val() is null or undefined, if it is then we return false (invalid).  This allows the require function to work with select boxes with a disabled selected value, such as:
<select id="month">
<option disabled="" selected="" value="">Month</option>
<option value="1">January</option>
...
</select>

That way we can just do $(#month).require() like a text box.  
-James